Pass node and volume name prefixes to tenks_update_state
node_name_prefix (default tk0) can be used to set a custom prefix for the node
names.

vol_name_prefix (default vol) can be used to set a custom prefix for the volume
names.

Support already existed for these in the action plugin, but the variables were
not passed through.

Change-Id: If968707285f9a8450bf16ca7a4956de6164986de
TrivialFix