Support ocata release of ironic
In the Ocata release, the physical_network field of ports did not exist.
The safest thing to do is to simply ignore it in this case.

It is also possible for the local_link_connection fields to not have switch_id,
switch_info or port_id attributes, so we now use .get(attr) to avoid an
attribute lookup error.

Change-Id: Ib4829dbd89fcedc121f493d6af76702421b399f7
TrivialFix