Ensure veth pairs are uniquely named
We name veth pairs using up to the first 6 characters from the nodes names. If
the node name prefixes are longer than this then the node index will not be
included, and the interface names will not be unique.

e.g. prefix compute, 2 nodes: compute0 & compute1

Link names are all p-comput-0-tap

Note that the 0 here is the index of the physnet, node the node.

This patch includes any trailing digits to keep link names unique, e.g.
p-compu0-0-tap.

Change-Id: I35c3da1d00030d8a270ac1a09e88e22098594f20
Story: 2007431
Task: 39058