Update TOX/UPPER_CONSTRAINTS_FILE for stable/1.0
Update the URL to the upper-constraints file to point to the redirect
rule on releases.openstack.org so that anyone working on this branch
will switch to the correct upper-constraints list automatically when
the requirements repository branches.

Until the requirements repository has as stable/1.0 branch, tests will
continue to use the upper-constraints list on master.

Change-Id: Ic0ab1f71920bd25246c301d095cd3000c8255dcc