Test all supported python version
As tenks is branchless which means same
master code is run on stable branches too, let's test it
with all the python version supported in those stable branches.

For that, we need to explicitly add the python jobs instead of using
the release specific job template. This way, we will explicitly know
that we are testing all supported python versions and do not remove the
older py version testing when the new release template bumps the min python
version like done in Yoga.

In Yoga testing, we are making py3.9 job to run as voting, so
adding py3.9 job as voting.

Change-Id: I1e9bbc75703f94acda0bddf0a2e43e835eb80cf0