Preserve assigned MAC on subsequent invocations
Tenks/libvirt was not reusing the MAC address assigned when invoked
repeatedly, which was leading to errors and confusion when extending
an existing Tenks deployment.

Check Ansible facts for a pre-existing interface and reuse the
assigned MAC if one is found.  Any existing MAC is added to the
data structure returned by set_libvirt_interfaces.

Change-Id: If45516d06d41992e7142ca8071ddcb84160a13e5